Refactor of GitHub JWT entities.
